@@ -4,25 +4,25 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.7.0",
-    "@craco/craco": "^6.3.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.10.1",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
+    "@craco/craco": "^6.4.3",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.0.3",
+    "@types/node": "^16.11.13",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
     "@types/react-lazyload": "^3.1.1",
-    "@types/react-router-dom": "^5.3.0",
-    "antd": "^4.16.13",
+    "@types/react-router-dom": "^5.3.2",
+    "antd": "^4.17.3",
     "craco-less": "^1.20.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-lazyload": "^3.2.0",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.1.1",
     "react-scripts": "4.0.3",
-    "typescript": "^4.4.3",
-    "web-vitals": "^2.1.0"
+    "typescript": "^4.5.4",
+    "web-vitals": "^2.1.2"
   },
   "scripts": {
     "start": "craco start",
